+++ packages/data-transforms/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@globalfishingwatch/data-transforms",
+  "version": "1.0.0",
+  "description": "Utils to decode various data formats",
+  "main": "dist/index.js",
+  "license": "MIT",
+  "scripts": {
+    "start": "rollup -c ./rollup.config.js -w",
+    "build": "yarn clean && cross-env NODE_ENV=production rollup -c rollup.config.js",
+    "clean": "rimraf dist/*",
+    "prepare": "yarn build",
+    "prepublishOnly": "yarn build"
+  }
+}
