+++ packages/data-transforms/rollup.config.js
@@ -1,0 +1,3 @@
+import rollup from '../../config/rollup.config'
+
+export default rollup()
