+++ packages/data-transforms/src/trackValueArrayToSegments/types.ts
@@ -1,0 +1,12 @@
+export enum Field {
+  lonlat = 'lonlat',
+  longitude = 'longitude',
+  latitude = 'latitude',
+  timestamp = 'timestamp',
+  fishing = 'fishing',
+  speed = 'speed',
+  course = 'course',
+}
+
+export type Point = Partial<Record<Field, number | null>>
+export type Segment = Point[]
