+++ packages/data-transforms/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../config/tsconfig.json",
+  "include": ["src/"],
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist/",
+    "typeRoots": ["../../node_modules/@types"]
+  }
+}
