+++ packages/layer-composer/src/index.ts
@@ -1,5 +1,4 @@
 export { default, DEFAULT_CONFIG } from './layer-composer'
 export { default as sort, convertLegacyGroups } from './transforms/sort/sort'
 export { default as getInteractiveLayerIds } from './transforms/getInteractiveLayerIds'
-export { default as trackSegments, TRACK_FIELDS } from './transforms/track-segments/track-segments'
 export * as Generators from './generators/types'
