--- packages/layer-composer/src/transforms/track-segments/index.ts
@@ -1,1 +1,0 @@
-export * from './track-segments'
